Rename ApiActionRetryDecider (#255)
