+++ tests/Unit/Services/RemoteRequestRetryDeciderTest.php
@@ -6,13 +6,13 @@ namespace App\Tests\Unit\Services;
 
 use App\Model\MachineProviderActionInterface;
 use App\Model\ProviderInterface;
-use App\Services\ApiActionRetryDecider;
-use App\Services\ApiActionRetryDecider\DigitalOcean\DigitalOceanApiActionRetryDecider;
+use App\Services\RemoteRequestRetryDecider;
+use App\Services\RemoteRequestRetryDecider\DigitalOcean\DigitalOceanRemoteRequestRetryDecider;
 use DigitalOceanV2\Exception\ApiLimitExceededException;
 use DigitalOceanV2\Exception\InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 
-class ApiActionRetryDeciderTest extends TestCase
+class RemoteRequestRetryDeciderTest extends TestCase
 {
     /**
      * @dataProvider decideDataProvider
@@ -21,7 +21,7 @@ class ApiActionRetryDeciderTest extends TestCase
      * @param MachineProviderActionInterface::ACTION_* $action
      */
     public function testDecide(
-        ApiActionRetryDecider $decider,
+        RemoteRequestRetryDecider $decider,
         string $provider,
         string $action,
         int $retryCount,
@@ -38,7 +38,7 @@ class ApiActionRetryDeciderTest extends TestCase
     {
         return [
             'no deciders' => [
-                'decider' => new ApiActionRetryDecider([], []),
+                'decider' => new RemoteRequestRetryDecider([], []),
                 'provider' => ProviderInterface::NAME_DIGITALOCEAN,
                 'action' => MachineProviderActionInterface::ACTION_GET,
                 'retryCount' => 0,
@@ -46,9 +46,9 @@ class ApiActionRetryDeciderTest extends TestCase
                 'expectedDecision' => false,
             ],
             'has decider, decider: false, retry limit not reached' => [
-                'decider' => new ApiActionRetryDecider(
+                'decider' => new RemoteRequestRetryDecider(
                     [
-                        new DigitalOceanApiActionRetryDecider(),
+                        new DigitalOceanRemoteRequestRetryDecider(),
                     ],
                     [
                         MachineProviderActionInterface::ACTION_GET => 10,
@@ -61,9 +61,9 @@ class ApiActionRetryDeciderTest extends TestCase
                 'expectedDecision' => false,
             ],
             'has decider, decider: true, retry limit not reached' => [
-                'decider' => new ApiActionRetryDecider(
+                'decider' => new RemoteRequestRetryDecider(
                     [
-                        new DigitalOceanApiActionRetryDecider(),
+                        new DigitalOceanRemoteRequestRetryDecider(),
                     ],
                     [
                         MachineProviderActionInterface::ACTION_GET => 10,
@@ -76,9 +76,9 @@ class ApiActionRetryDeciderTest extends TestCase
                 'expectedDecision' => true,
             ],
             'has decider, decider: true, retry limit reached' => [
-                'decider' => new ApiActionRetryDecider(
+                'decider' => new RemoteRequestRetryDecider(
                     [
-                        new DigitalOceanApiActionRetryDecider(),
+                        new DigitalOceanRemoteRequestRetryDecider(),
                     ],
                     [
                         MachineProviderActionInterface::ACTION_GET => 1,
